# columns/filter: Add support for time.Duration