Fixes: https://github.com/inspektor-gadget/inspektor-gadget/issues/2752

### before the PR
```bash
# nslookup inspektor-gadget.io
→ go run -exec 'sudo' ./cmd/ig trace dns -o columns=+latency --host -F 'latency:>10ms'
RUNTIME.CONTAINERNAME                                 PID        TID        COMM             QR TYPE      QTYPE      NAME                                                 RCODE                        NUMANSW… LATENCY             
                                                      2697972    2697973    isc-net-0000     R  HOST      A          inspektor-gadget.io.                                 No Error                     2        394.511µs           
                                                      2697972    2697973    isc-net-0000     R  HOST      AAAA       inspektor-gadget.io.                                 No Error                     2        292.461µs           
                                                      933        933        systemd-resolve  R  HOST      A          copilot-telemetry.githubusercontent.com.             No Error                     3        11.544263ms         
                                                      2615202    2615221    copilot-agent-l  R  HOST      A          copilot-telemetry.githubusercontent.com.             No Error                     3        12.06946ms          
                                                      933        933        systemd-resolve  R  HOST      AAAA       copilot-telemetry.githubusercontent.com.             No Error                     2        11.882737ms         
                                                      933        933        systemd-resolve  R  HOST      AAAA       glb-db52c2cf8be544.github.com.                       No Error                     0        8.692058ms          
                                                      2615202    2615221    copilot-agent-l  R  HOST      AAAA       copilot-telemetry.githubusercontent.com.             No Error                     2        21.468557ms 
```

## after the PR
```bash
# nslookup  ebpf.io
→ go run -exec 'sudo' ./cmd/ig trace dns -o columns=+latency --host -F 'latency:>10ms'
RUNTIME.CONTAINERNAME                                 PID        TID        COMM             QR TYPE      QTYPE      NAME                                                 RCODE                        NUMANSW… LATENCY                    
                                                      933        933        systemd-resolve  R  HOST      A          ebpf.io.                                             No Error                     1        46.93168ms          
                                                      2698986    2698988    isc-net-0000     R  HOST      A          ebpf.io.                                             No Error                     1        47.336763ms         
                                                      933        933        systemd-resolve  R  HOST      A          ssl.gstatic.com.                                     No Error                     1        13.134705ms         
                                                      2624925    2624930    Chrome_ChildIOT  R  HOST      A          ssl.gstatic.com.                                     No Error                     1        13.765444ms         
                                                      933        933        systemd-resolve  R  HOST      AAAA       ebpf.io.                                             No Error                     0        47.554799ms         
                                                      2698986    2698988    isc-net-0000     R  HOST      AAAA       ebpf.io.                                             No Error                     0        47.986683ms  
```